Update code to add parameter sliders upon simulation start without user intervention. Fix bug handling null nodes when plotting species
When simulation is started, up to 10 parameter sliders are automatically created. If more than 10 parameters then user has to pick which ones. Bug handling null nodes in reactions. Was not checking if stoich coefficient for reaction species exists, which does not for null node. Tried to add nil value to string for rate equation