Fix button hint (tooltip) issue where hint popup stayed visible after mouse focus changes
When update Bootstrap from 4 to 5 this issue appeared. The popper version that works with Bootstrap 5 seems to have changed the tooltip default behavior (or there is a bug).  The issue was the hint popup with mouseover, as expected, but does not go away with loss of focus. pretty soon all the button hints are visible, covering up the windows. Going back to bootstrap 4, with associated popper.js fixed the issue.