add code to handle change in step size without changing chart time scale
Now if user changes time step, the x axis scale does not change unless the total number of points to plot goes above 1000. Previously it was 100. Also if the number of points to plot goes below 50 then the default value of 100 is used to calculate the x axis time scale.